percpu
Struct EXAMPLE_PERCPU_DATA_WRAPPER

pub struct EXAMPLE_PERCPU_DATA_WRAPPER {}

Wrapper struct for the per-CPU data EXAMPLE_PERCPU_DATA

Implementations

impl EXAMPLE_PERCPU_DATA_WRAPPER

pub fn offset(&self) -> usize

Returns the offset relative to the per-CPU data area base.

pub unsafe fn current_ptr(&self) -> *const usize

Returns the raw pointer of this per-CPU static variable on the current CPU.

Safety

Caller must ensure that preemption is disabled on the current CPU.

pub unsafe fn current_ref_raw(&self) -> &usize

Returns the reference of the per-CPU static variable on the current CPU.

Safety

Caller must ensure that preemption is disabled on the current CPU.

pub unsafe fn current_ref_mut_raw(&self) -> &mut usize

Returns the mutable reference of the per-CPU static variable on the current CPU.

Safety

Caller must ensure that preemption is disabled on the current CPU.

pub fn with_current<F, T>(&self, f: F) -> T
where F: FnOnce(&mut usize) -> T,

Manipulate the per-CPU data on the current CPU in the given closure. Preemption will be disabled during the call.

pub unsafe fn remote_ptr(&self, cpu_id: usize) -> *const usize

Returns the raw pointer of this per-CPU static variable on the given CPU.

Safety

Caller must ensure that

the CPU ID is valid, and

data races will not happen.

pub unsafe fn remote_ref_raw(&self, cpu_id: usize) -> &usize

Returns the reference of the per-CPU static variable on the given CPU.

Safety

Caller must ensure that

the CPU ID is valid, and

data races will not happen.

pub unsafe fn remote_ref_mut_raw(&self, cpu_id: usize) -> &mut usize

Returns the mutable reference of the per-CPU static variable on the given CPU.

Safety

Caller must ensure that

the CPU ID is valid, and

data races will not happen.

pub unsafe fn read_current_raw(&self) -> usize

Returns the value of the per-CPU static variable on the current CPU.

Safety

Caller must ensure that preemption is disabled on the current CPU.

pub unsafe fn write_current_raw(&self, val: usize)

Set the value of the per-CPU static variable on the current CPU.

Safety

Caller must ensure that preemption is disabled on the current CPU.

pub fn read_current(&self) -> usize

Returns the value of the per-CPU static variable on the current CPU. Preemption will be disabled during the call.

pub fn write_current(&self, val: usize)

Set the value of the per-CPU static variable on the current CPU. Preemption will be disabled during the call.